% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_filtered_results.R
\name{bin_filtered_results}
\alias{bin_filtered_results}
\title{bin_filtered_results}
\usage{
bin_filtered_results(filtered_edna, input_destination, data_location)
}
\arguments{
\item{destination}{}
}
\value{

}
\description{
bin_filtered_results
}
