% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_directories.R
\name{check_directories}
\alias{check_directories}
\title{check_directories}
\usage{
check_directories(input_location, data_location)
}
\arguments{
\item{data_location}{}
}
\value{

}
\description{
check_directories
}
