% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_updates.R
\name{check_updated_inputs}
\alias{check_updated_inputs}
\title{check updated inputs}
\usage{
check_updated_inputs(destination)
}
\arguments{
\item{destination}{}
}
\value{

}
\description{
check updated inputs
}
