% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_updates.R
\name{check_updated_publist}
\alias{check_updated_publist}
\title{check updated publist}
\usage{
check_updated_publist(input_destination)
}
\arguments{
\item{destination}{}
}
\value{

}
\description{
check updated publist
}
