% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_edna_results.R
\name{filter_edna_results}
\alias{filter_edna_results}
\title{filter_edna_results}
\usage{
filter_edna_results(results, destination)
}
\arguments{
\item{results}{}
}
\value{

}
\description{
filter_edna_results
}
