% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_crossref_edna.R
\name{query_crossref_edna}
\alias{query_crossref_edna}
\title{query_crossref_edna}
\usage{
query_crossref_edna(backup, email)
}
\arguments{
\item{email}{}
}
\value{

}
\description{
query_crossref_edna
}
