% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_filtered_results.R
\name{save_filtered_results}
\alias{save_filtered_results}
\title{save_filtered_results}
\usage{
save_filtered_results(filtered_results, input_destination)
}
\arguments{
\item{filtered_results}{}
}
\value{

}
\description{
save_filtered_results
}
