% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_destinations.R
\name{set_backup_destination}
\alias{set_backup_destination}
\title{set_backup_destination}
\usage{
set_backup_destination(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
set_backup_destination
}
