% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_edna_knowledge_center.R
\name{update_edna_knowledge_center}
\alias{update_edna_knowledge_center}
\title{update_edna_knowledge_center}
\usage{
update_edna_knowledge_center(destination, backup, email)
}
\arguments{
\item{email}{}
}
\value{

}
\description{
update_edna_knowledge_center
}
